+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ###### Backwards compatible bug fixes
 - #69 - Failed requests should throw an error, not just return it
+- #70 - Make `params` and `params.query` optional
 
 ##### 0.9.0 - 22 May 2014